style(fsrepo): rename to counter.Openers
