+++ repo/fsrepo/counter/openers.go
@@ -1,17 +1,17 @@
-package fsrepo
+package counter
 
 import "path"
 
 // TODO this could be made into something more generic.
 
-type Counter struct {
+type Openers struct {
 	// repos maps repo paths to the number of openers holding an FSRepo handle
 	// to it
 	repos map[string]int
 }
 
-func NewCounter() *Counter {
-	return &Counter{
+func NewOpenersCounter() *Openers {
+	return &Openers{
 		repos: make(map[string]int),
 	}
 }
@@ -19,13 +19,13 @@ func NewCounter() *Counter {
 // NumOpeners returns the number of FSRepos holding a handle to the repo at
 // this path. This method is not thread-safe. The caller must have this object
 // locked.
-func (l *Counter) NumOpeners(repoPath string) int {
+func (l *Openers) NumOpeners(repoPath string) int {
 	return l.repos[key(repoPath)]
 }
 
 // AddOpener messages that an FSRepo holds a handle to the repo at this path.
 // This method is not thread-safe. The caller must have this object locked.
-func (l *Counter) AddOpener(repoPath string) error {
+func (l *Openers) AddOpener(repoPath string) error {
 	l.repos[key(repoPath)]++
 	return nil
 }
@@ -33,7 +33,7 @@ func (l *Counter) AddOpener(repoPath string) error {
 // RemoveOpener messgaes that an FSRepo no longer holds a handle to the repo at
 // this path. This method is not thread-safe. The caller must have this object
 // locked.
-func (l *Counter) RemoveOpener(repoPath string) error {
+func (l *Openers) RemoveOpener(repoPath string) error {
 	l.repos[key(repoPath)]--
 	return nil
 }
